// packages/collaboration/src/controllers/config.schema.ts
/**
 * Copyright 2023-present DreamNum Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import type { IWorkbookData } from '@univerjs/core';
import type { MenuConfig } from '@univerjs/ui';

export const COLLABORATION_PLUGIN_CONFIG_KEY = 'collaboration.config';

export const configSymbol = Symbol(COLLABORATION_PLUGIN_CONFIG_KEY);

export interface ICollaborator {
    userId: string;
    userName?: string;
    avatar?: string;
}

export interface IUniverCollaborationConfig {
    menu?: MenuConfig;
    contentId?: string;
    collaborationUrl?: string;
    collaborator?: ICollaborator;
    data?: IWorkbookData | undefined;
}

export const defaultPluginConfig: IUniverCollaborationConfig = {};